# demo-ui/src/app/services/models/student-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StudentRequest {
  departementId: number;
  email: string;
  id?: number;
  name: string;
}