# demo-ui/src/app/services/models/student-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StudentResponse {
  departementName?: string;
  email?: string;
  id?: number;
  name?: string;
}