# demo-ui/src/app/services/models/update-student-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateStudentRequest {
  departementId?: number;
  email?: string;
  id?: number;
  name?: string;
}